+++ llm_ffw/__init__.py
@@ -8,6 +8,7 @@ from .capabilities import (
     SecretCatalogCapability,
     UnsafeURLCapability,
     PaymentCardCapability,
+    PrivateKeyCapability,
 )
 from .banned_substring_catalog import (
     BannedSubstring,
@@ -27,6 +28,7 @@ from .inspection import (
 from .json_output import JSONOutputConfig
 from .unsafe_url import UnsafeURLConfig
 from .payment_card import PaymentCardConfig
+from .private_key import PrivateKeyConfig
 from .manager import (
     FirewallManagerState,
     FirewallReloadError,
@@ -59,6 +61,7 @@ from .rules.banned_substrings import BannedSubstringsRule
 from .rules.json_output import JSONOutputRule
 from .rules.unsafe_url import UnsafeURLRule
 from .rules.payment_card import PaymentCardRule
+from .rules.private_key import PrivateKeyRule
 
 __all__ = [
     "Action",
@@ -98,6 +101,9 @@ __all__ = [
     "PaymentCardCapability",
     "PaymentCardConfig",
     "PaymentCardRule",
+    "PrivateKeyCapability",
+    "PrivateKeyConfig",
+    "PrivateKeyRule",
     "Scanner",
     "ScannerConfig",
     "ScanScope",