+++ llm_ffw/rules/__init__.py
@@ -5,6 +5,7 @@ from .banned_substrings import BannedSubstringsRule
 from .invisible_characters import InvisibleCharactersRule
 from .json_output import JSONOutputRule
 from .payment_card import PaymentCardRule
+from .private_key import PrivateKeyRule
 from .secrets import SecretsRule
 from .unsafe_url import UnsafeURLRule
 
@@ -13,6 +14,7 @@ __all__ = [
     "InvisibleCharactersRule",
     "JSONOutputRule",
     "PaymentCardRule",
+    "PrivateKeyRule",
     "Rule",
     "RuleMatch",
     "SecretsRule",